+++ src/data/genders.js
@@ -1,0 +1,15 @@
+export default [
+  'agender',
+  'androgyne',
+  'bigender',
+  'genderfluid',
+  'nonconforming',
+  'questioning',
+  'variant',
+  'genderqueer',
+  'neutrois',
+  'pangender',
+  'transgender',
+  'transsexual',
+  'other',
+]